data/requirements.txt
nats-py
deltalake
s3fs
boto3